+++ actrec/operators/__init__.py
@@ -1,0 +1,21 @@
+from .categories import (
+    AR_OT_category_add,
+    AR_OT_category_edit,
+    AR_OT_category_apply_visibility,
+    AR_OT_category_delete_visibility,
+    AR_OT_category_delete
+)
+
+from .globals import (
+    AR_OT_gloabal_recategorize_action
+)
+
+from .preferences import(
+    AR_OT_preferences_directory_selector,
+    AR_OT_preferences_recover_directory
+)
+
+from .shared import (
+    AR_OT_check_ctrl,
+    AR_OT_open_url
+)
