+++ actrec/properties/__init__.py
@@ -1,0 +1,17 @@
+from .categories import (
+    AR_selected_category,
+    AR_categories
+)
+
+from .globals import (
+    AR_global_actions_enum
+)
+
+from .preferences import (
+    AR_macro,
+    AR_action
+)
+
+from .shared import (
+    id_system
+)
